services/ad-review-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.2
httpx==0.27.0
chardet==5.2.0
python-multipart==0.0.9
openpyxl==3.1.5
sse-starlette==2.0.0
python-docx==1.1.2
PyPDF2==3.0.1
xlrd==2.0.1
msoffcrypto-tool==5.4.2
